+++ fortnite_wiki/fortnite_wiki/Viewcontrollers/DetailsWeaponViewController.swift
@@ -10,7 +10,7 @@
 import UIKit
 import ChameleonFramework
 
-class DetailsViewController: UIViewController {
+class DetailsWeaponViewController: UIViewController {
     
     let colors = BackgroundColors()